refactor(Table): move fillColumn into hook
